Fixed compiler warnings and turned on -Werror

+++ source/sclpl/main.c
@@ -6,10 +6,12 @@
 #include "parser.h"
 #include "lexer.h"
 #include "pprint.h"
+#include "codegen.h"
 
 /* Command Line Options
  *****************************************************************************/
 const char Usage[] = "Usage: sclpl [OPTION]... MODE [FILE]...\n";
+
 opts_cfg_t Options_Config[] = {
     {"tokens",    false, "mode",    "Emit the token output of lexical analysis for the given file"},
     {"ast",       false, "mode",    "Emit the abstract syntax tree for the given file"},
@@ -42,7 +44,7 @@ void print_usage(void) {
 
     /* Print the usage and option list */
     puts(Usage);
-    size_t padding = sz + 4 + ((opts_have_args) ? 4 : 0);
+    int padding = sz + 4 + ((opts_have_args) ? 4 : 0);
     char*  buffer  = (char*)malloc(padding+1);
     opts = &Options_Config[0];
     while (NULL != opts->name) {
